+++ packages/core/src/query.ts
@@ -1,11 +1,13 @@
 import { AsyncLocalStorage } from 'node:async_hooks';
 import { AdapterBase } from './adapter';
+import { EmptyObject } from './utils';
 
 // query metadata that is stored only on TS side, not available in runtime
 export type QueryMetaBase = {
   as?: string;
   hasSelect?: true;
   hasWhere?: true;
+  defaults: EmptyObject;
 };
 
 export type QueryInternal = {